All printf's of the world return int - this warning has driven me crazy   for months.  Sorry for the cast, but this one's quite harmless.
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@61710 13f79535-47bb-0310-9956-ffa450edef68

+++ network_io/unix/inet_ntop.c
@@ -102,7 +102,7 @@ inet_ntop4(const unsigned char *src, char *dst, apr_size_t size)
 	static const char fmt[] = "%u.%u.%u.%u";
 	char tmp[sizeof "255.255.255.255"];
 
-	if (apr_snprintf(tmp, sizeof tmp, fmt, src[0], src[1], src[2], src[3]) > size) {
+	if (apr_snprintf(tmp, sizeof tmp, fmt, src[0], src[1], src[2], src[3]) > (int)size) {
 		errno = ENOSPC;
 		return (NULL);
 	}
